backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
supabase==2.3.0
langchain==0.0.350
langchain-google-genai==0.0.5
pydantic>=2.0
python-dotenv==1.0.0
pytest==7.4.3
hypothesis==6.92.1
python-multipart==0.0.6
